refactor: change paths into user module
