+++ src/user/user.module.ts
@@ -1,7 +1,7 @@
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
 import { Module } from '@nestjs/common';
-import { DbModule } from './db/db.module';
+import { DbModule } from '../db/db.module';
 
 @Module({
   imports: [DbModule],
